feat: 오픈소스 버전 테스트 코드 분리 [#16]
