+++ src/test/kotlin/com/osscube/api/domain/service/open_source_version/AddNewVersionTest.kt
@@ -1,4 +1,4 @@
-package com.osscube.api.domain.service
+package com.osscube.api.domain.service.open_source_version
 
 import com.osscube.api.config.TestContainers
 import com.osscube.api.domain.dto.OpenSourceVersionAddRequestDto
@@ -8,9 +8,9 @@ import com.osscube.api.domain.model.entity.OpenSource
 import com.osscube.api.domain.model.entity.OpenSourceVersion
 import com.osscube.api.domain.model.repository.OpenSourceRepository
 import com.osscube.api.domain.model.repository.OpenSourceVersionRepository
+import com.osscube.api.domain.service.OpenSourceVersionService
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
-import org.assertj.core.api.Assertions.tuple
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
@@ -20,7 +20,7 @@ import org.testcontainers.junit.jupiter.Testcontainers
 
 @SpringBootTest
 @Testcontainers
-class OpenSourceVersionServiceTest : TestContainers() {
+class AddNewVersionTest : TestContainers() {
     @Autowired
     private lateinit var openSourceVersionService: OpenSourceVersionService
 
@@ -84,44 +84,5 @@ class OpenSourceVersionServiceTest : TestContainers() {
         val requestDto = OpenSourceVersionAddRequestDto(version, sourceUrl)
         assertThatThrownBy { openSourceVersionService.addNewVersion(clientId, requestDto) }
             .isInstanceOf(OpenSourceVersionAlreadyExistsException::class.java)
-    }
-
-    @DisplayName("openSourceId에 해당하는 오픈소스의 모든 버전 조회하기")
-    @Test
-    fun getVersions() {
-        // given
-        val openSource = OpenSource("JSON-java", "https://github.com/stleary/JSON-java")
-        openSourceRepository.save(openSource)
-
-        val givenVersions = listOf(
-            OpenSourceVersion(openSource, "20240303", "https://github.com/stleary/JSON-java/archive/refs/tags/20240303.tar.gz"),
-            OpenSourceVersion(openSource, "20240205", "https://github.com/stleary/JSON-java/archive/refs/tags/20240205.tar.gz"),
-            OpenSourceVersion(openSource, "20231013", null)
-        )
-        openSourceVersionRepository.saveAll(givenVersions)
-
-        // when
-        val openSourceVersions = openSourceVersionService.getVersions(openSource.clientId)
-
-        // then
-        openSourceVersions.forEach { version -> assertThat(version.id).hasSize(36) }
-        assertThat(openSourceVersions)
-            .hasSize(3)
-            .extracting("version", "sourceUrl")
-            .contains(
-                tuple("20240303", "https://github.com/stleary/JSON-java/archive/refs/tags/20240303.tar.gz"),
-                tuple("20240205", "https://github.com/stleary/JSON-java/archive/refs/tags/20240205.tar.gz"),
-                tuple("20231013", null)
-            )
-    }
-
-    @DisplayName("오픈소스가 존재하지 않으면 버전을 조회할 수 없다.")
-    @Test
-    fun cannotGetVersionsIfOpenSourceNotExists() {
-        // given
-
-        // when // then
-        assertThatThrownBy { openSourceVersionService.getVersions("invalid open source id") }
-            .isInstanceOf(OpenSourceNotFoundException::class.java)
     }
 }
